pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Fedor Kuznetsov <kuzneczov-f@susanoo.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rasterio = "^1.3.9"
matplotlib = "^3.8.3"
geopandas = "^0.14.3"
flask = "^3.0.2"
dash = "^2.15.0"
plotly = "^5.19.0"
folium = "^0.16.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
redis = "^5.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
